Create utils, handle error
