+++ apps/src/sketchlab/types.ts
@@ -1,0 +1,19 @@
+import {ExcalidrawElement} from '@excalidraw/excalidraw/types/element/types';
+import {AppState, BinaryFiles} from '@excalidraw/excalidraw/types/types';
+
+import {
+  ExcalidrawSourceWithExternalFiles,
+  ProjectSources,
+  SketchlabExternalFiles,
+} from '@cdo/apps/lab2/types';
+
+export interface SerializedExcalidrawState {
+  elements: ExcalidrawElement[];
+  appState: AppState;
+  files: BinaryFiles;
+  externalFiles?: SketchlabExternalFiles;
+}
+
+export interface SketchlabSources extends ProjectSources {
+  source: ExcalidrawSourceWithExternalFiles;
+}
