+++ apps/src/sketchlab/utils/uploadExternalFiles.ts
@@ -1,0 +1,70 @@
+import {SketchlabProjectFile} from '@cdo/apps/lab2/types';
+
+import {SketchlabSources, SerializedExcalidrawState} from '../types';
+
+import uploadBase64ToUrl from './uploadBase64ToUrl';
+
+// TO DO: add Jira to handle bad extensions.
+const MIME_TO_EXT = {
+  'image/svg+xml': 'svg',
+  'image/png': 'png',
+  'image/jpeg': 'jpg',
+  'image/gif': 'gif',
+  'image/webp': 'webp',
+  'image/bmp': 'bmp',
+  'image/x-icon': 'ico',
+  'image/avif': 'avif',
+  'image/jfif': 'jfif',
+  'application/octet-stream': 'bin',
+};
+
+const uploadExternalFiles = (
+  source: ExcalidrawSourceWithExternalFiles,
+  serializedData: SerializedExcalidrawState,
+  filesBeingUploadedRef: React.MutableRefObject<Set<string>>,
+  channelId: string,
+  updateSources: (newSources: SketchlabSources, forceSave?: boolean) => void
+) => {
+  const savedFileIds = Object.keys(source.externalFiles || {});
+  const excalidrawFileIds = Object.keys(serializedData.files || {});
+  const difference = excalidrawFileIds.filter(
+    id => !savedFileIds.includes(id) && !filesBeingUploadedRef.current.has(id)
+  );
+
+  if (difference.length && serializedData.files) {
+    difference.map(async fileId => {
+      filesBeingUploadedRef.current.add(fileId);
+
+      const newFile = serializedData.files[fileId];
+      const extension = MIME_TO_EXT[newFile.mimeType];
+      const externalUrl = `/v3/assets/${channelId}/${fileId}.${extension}`;
+      const newExternalFile: SketchlabProjectFile = {
+        id: fileId,
+        url: externalUrl,
+      };
+
+      try {
+        await uploadBase64ToUrl(newFile.dataURL, externalUrl, newFile.mimeType);
+      } catch {
+        // If an upload fails, still add an entry to externalFiles
+        // so we don't reattempt the upload repeatedly.
+        // Longer term work to handle failed uploads tracked here:
+        // https://codedotorg.atlassian.net/browse/AFL-345
+        newExternalFile.uploadFailed = true;
+      }
+
+      updateSources({
+        source: {
+          ...source,
+          externalFiles: {
+            ...source.externalFiles,
+            [fileId]: newExternalFile,
+          },
+        },
+      });
+      filesBeingUploadedRef.current.delete(fileId);
+    });
+  }
+};
+
+export default uploadExternalFiles;
